Engraving: move hairpin commands from Score into EditHairpin
