+++ src/engraving/editing/edit.cpp
@@ -24,7 +24,6 @@
 #include <set>
 
 #include "../dom/accidental.h"
-#include "../dom/anchors.h"
 #include "../dom/arpeggio.h"
 #include "../dom/articulation.h"
 #include "../dom/barline.h"
@@ -45,7 +44,6 @@
 #include "../dom/fret.h"
 #include "../dom/glissando.h"
 #include "../dom/guitarbend.h"
-#include "../dom/hairpin.h"
 #include "../dom/hammeronpulloff.h"
 #include "../dom/harmony.h"
 #include "../dom/harppedaldiagram.h"
@@ -2515,197 +2513,6 @@ void Score::cmdFullMeasureRest()
     } else {
         deselectAll();
     }
-}
-
-std::vector<Hairpin*> Score::addHairpins(HairpinType type)
-{
-    std::vector<Hairpin*> hairpins;
-
-    // add hairpin on each staff if possible
-    if (selection().isRange() && selection().staffStart() != selection().staffEnd() - 1) {
-        for (staff_idx_t staffIdx = selection().staffStart(); staffIdx < selection().staffEnd(); ++staffIdx) {
-            ChordRest* cr1 = selection().firstChordRest(staffIdx * VOICES);
-            ChordRest* cr2 = selection().lastChordRest(staffIdx * VOICES);
-            Hairpin* h = cr1 ? addHairpin(type, cr1, cr2)
-                         : addHairpin(type, selection().tickStart(), selection().tickEnd(), staff2track(staffIdx));
-            if (h) {
-                hairpins.push_back(h);
-            }
-        }
-    } else {
-        // for single staff range selection, or single selection,
-        // find start & end elements elements
-        ChordRest* cr1 = nullptr;
-        ChordRest* cr2 = nullptr;
-        getSelectedStartEndChordRests(cr1, cr2);
-        Hairpin* h = addHairpin(type, cr1, cr2);
-        if (h) {
-            hairpins.push_back(h);
-        }
-    }
-
-    for (Hairpin* hairpin : hairpins) {
-        hairpin->setInitialTrackAndVoiceAssignment(hairpin->track(), /*curVoiceOnlyOverride*/ false);
-    }
-
-    return hairpins;
-}
-
-Hairpin* Score::addHairpin(HairpinType type, ChordRest* cr1, ChordRest* cr2)
-{
-    if (!cr1) {
-        return nullptr;
-    }
-
-    Hairpin* hairpin = Factory::createHairpin(this->dummy()->segment());
-    hairpin->setHairpinType(type);
-    if (type == HairpinType::CRESC_LINE) {
-        hairpin->setBeginText(u"cresc.");
-        hairpin->setContinueText(u"(cresc.)");
-    } else if (type == HairpinType::DIM_LINE) {
-        hairpin->setBeginText(u"dim.");
-        hairpin->setContinueText(u"(dim.)");
-    }
-
-    addHairpin(hairpin, cr1, cr2);
-
-    return hairpin;
-}
-
-Hairpin* Score::addHairpin(HairpinType type, Fraction sTick, Fraction eTick, track_idx_t track)
-{
-    Hairpin* hairpin = Factory::createHairpin(this->dummy()->segment());
-    hairpin->setHairpinType(type);
-    if (type == HairpinType::CRESC_LINE) {
-        hairpin->setBeginText(u"cresc.");
-        hairpin->setContinueText(u"(cresc.)");
-    } else if (type == HairpinType::DIM_LINE) {
-        hairpin->setBeginText(u"dim.");
-        hairpin->setContinueText(u"(dim.)");
-    }
-
-    hairpin->setTrack(track);
-    hairpin->setTick(sTick);
-    hairpin->setTick2(eTick);
-
-    undoAddElement(hairpin);
-
-    return hairpin;
-}
-
-void Score::addHairpin(Hairpin* hairpin, ChordRest* cr1, ChordRest* cr2)
-{
-    track_idx_t track = cr1->track();
-    Fraction startTick = cr1->tick();
-
-    Fraction endTick = startTick;
-    if (m_selection.isRange()) {
-        endTick = m_selection.tickEnd();
-    } else {
-        if (cr2 && cr2 != cr1) {
-            endTick = cr2->endTick();
-        } else {
-            endTick = cr1->isChord() ? toChord(cr1)->endTickIncludingTied() : cr1->endTick();
-        }
-        const Segment* startSegment = cr2 ? cr2->segment() : cr1->segment();
-        for (const Segment* segment = startSegment; segment && segment->tick() < endTick;
-             segment = segment->next1(SegmentType::Duration)) {
-            if (segment == startSegment) {
-                continue;
-            }
-            if (segment->findAnnotation(ElementType::DYNAMIC, track, track)) {
-                endTick = segment->tick();
-                break;
-            }
-        }
-    }
-
-    hairpin->setTrack(track);
-    hairpin->setTrack2(track);
-    hairpin->setTick(startTick);
-    hairpin->setTick2(endTick);
-
-    undoAddElement(hairpin);
-}
-
-void Score::addHairpinToDynamic(Hairpin* hairpin, Dynamic* dynamic)
-{
-    track_idx_t track = dynamic->track();
-    hairpin->setTrack(track);
-    hairpin->setTrack2(track);
-
-    hairpin->setTick(dynamic->tick());
-
-    Segment* dynamicSegment = dynamic->segment();
-
-    Chord* startChord = nullptr;
-    for (Segment* segment = dynamicSegment; segment; segment = segment->prev(SegmentType::ChordRest)) {
-        EngravingItem* element = segment->element(track);
-        if (element && element->isChord()) {
-            startChord = toChord(element);
-            break;
-        }
-    }
-
-    Fraction endTick = startChord ? startChord->endTickIncludingTied() : dynamic->segment()->measure()->endTick();
-
-    for (Segment* segment = dynamicSegment; segment && segment->tick() < endTick;
-         segment = segment->next1(SegmentType::Duration)) {
-        if (segment == dynamicSegment) {
-            continue;
-        }
-        if (segment->findAnnotation(ElementType::DYNAMIC, track, track)) {
-            endTick = segment->tick();
-            break;
-        }
-    }
-
-    hairpin->setTick2(endTick);
-
-    hairpin->setVoiceAssignment(dynamic->voiceAssignment());
-
-    undoAddElement(hairpin);
-}
-
-Hairpin* Score::addHairpinToDynamicOnGripDrag(Dynamic* dynamic, bool isLeftGrip, const PointF& pos)
-{
-    const track_idx_t track = dynamic->track();
-    staff_idx_t staffIndex = dynamic->staffIdx();
-    Segment* seg = nullptr;
-    constexpr double spacingFactor = 0.5;
-
-    // Ensure time tick segments are created
-    EditTimeTickAnchors::updateAnchors(dynamic);
-
-    // Find segment of type ChordRest or TimeTick near cursor postion
-    dragPosition(pos, &staffIndex, &seg, spacingFactor, /*allowTimeAnchor*/ true);
-
-    const bool hasValidTick = seg && (isLeftGrip
-                                      ? seg->tick() < dynamic->tick()
-                                      : seg->tick() > dynamic->tick());
-    if (!hasValidTick) {
-        return nullptr;
-    }
-
-    Hairpin* hairpin = Factory::createHairpin(dummy()->segment());
-    hairpin->setHairpinType(isLeftGrip ? HairpinType::DIM_HAIRPIN : HairpinType::CRESC_HAIRPIN);
-
-    hairpin->setTrack(track);
-    hairpin->setTrack2(track);
-
-    if (isLeftGrip) {
-        hairpin->setTick(seg->tick());
-        hairpin->setTick2(dynamic->tick());
-    } else {
-        hairpin->setTick(dynamic->tick());
-        hairpin->setTick2(seg->tick());
-    }
-
-    hairpin->setVoiceAssignment(dynamic->voiceAssignment());
-
-    undoAddElement(hairpin);
-
-    return hairpin;
 }
 
 //---------------------------------------------------------
