+++ src/notationscene/qml/MuseScore/NotationScene/elementpopups/dynamicpopupmodel.cpp
@@ -24,6 +24,8 @@
 
 #include "engraving/dom/dynamic.h"
 #include "engraving/dom/factory.h"
+#include "engraving/editing/edithairpin.h"
+#include "engraving/editing/transaction/transaction.h"
 #include "engraving/types/symnames.h"
 #include "engraving/types/typesconv.h"
 
@@ -207,7 +209,8 @@ void DynamicPopupModel::addHairpinToDynamic(ItemType itemType)
     beginCommand(TranslatableString("undoableAction", "Add hairpin"));
     Hairpin* hairpin = Factory::createHairpin(m_item->score()->dummy()->segment());
     hairpin->setHairpinType(hairpinType);
-    m_item->score()->addHairpinToDynamic(hairpin, dynamic);
+    Transaction& tx = m_item->score()->transactionManager()->currentOrDummyTransaction();
+    EditHairpin::addHairpinToDynamic(tx, m_item->score(), hairpin, dynamic);
     endCommand();
     interaction->selectAndStartEditIfNeeded(hairpin);
     updateNotation();
